modifications to improve memory usage during assignation in accumulation

+++ restructure/blob.py
@@ -361,11 +361,11 @@ def apply_model_area(video, blob, model_area, portraitSize):
     if model_area(blob.area): #Checks if area is compatible with the model area we built
         if video.animal_type == 'fish':
             portrait, blob._nose_coordinates, blob._head_coordinates = getPortrait(blob.bounding_box_image, blob.contour, blob.bounding_box_in_frame_coordinates, portraitSize)
-            blob._portrait = (portrait - np.mean(portrait))/np.std(portrait)
+            blob._portrait = ((portrait - np.mean(portrait))/np.std(portrait)).astype('float32')
             blob.bounding_box_image = None
         elif video.animal_type == 'fly':
             portrait, blob._extreme1_coordinates, blob._extreme2_coordinates = get_portrait_fly(video, blob.bounding_box_image, blob.pixels, blob.bounding_box_in_frame_coordinates, portraitSize)
-            blob._portrait = (portrait - np.mean(portrait))/np.std(portrait)
+            blob._portrait = ((portrait - np.mean(portrait))/np.std(portrait)).astype('float32')
             blob.bounding_box_image = None
 
 def apply_model_area_to_blobs_in_frame(video, blobs_in_frame, model_area, portraitSize):
